requirements.txt
# Minimal Python requirements for the na-radio webcam demo
# Minimal Python requirements for the na-radio webcam demo
# torch and torchvision are installed by the setup scripts/Dockerfile to match the selected runtime (CPU vs CUDA)
# They are deliberately removed to avoid accidental installation of mismatched GPU wheels
numpy<2; python_version < "3.13"
numpy>=2.0; python_version >= "3.13"
# If you plan to use a full GUI (cv2.imshow) and camera capture, install opencv-python
# On some distros / NixOS, you may prefer opencv-python-headless and a separate v4l package
opencv-python
matplotlib
timm
typing-extensions
einops
open_clip_torch==2.24.0
imageio
flask
gunicorn
werkzeug
transformers
# Optional but recommended when running RADIO/CLIP models
# open_clip_torch
# sentence-transformers
ultralytics
segment-anything
git+https://github.com/ultralytics/CLIP.git
# OpenYOLO3D dependencies
openmim
mmengine
mmcv>=2.0.0
mmdet>=3.0.0
mmyolo
pyyaml



